moved matmul to discrete operator
